Part 9 imaginary is done

+++ 10 Operator Overloading/1_complex_numbers_example_header.h
@@ -9,11 +9,10 @@ class Complex{
         float real;
         float ima;
     public:
-        Complex(int, int);
+        Complex(float, float);
         friend ostream& operator<<(ostream& , Complex &);
         Complex operator+(Complex);
         friend Complex operator-(Complex, Complex);
 };
-
 
 #endif
